database.py now calculates min_commits and adds it to the assignemnts table
