+++ database.py
@@ -1,8 +1,56 @@
 import sqlite3
-from os import listdir
+from os import listdir, chdir
+from os.path import exists, isdir
 from sys import exit
 
 DATABASE_NAME = 'data.sqlite3'
+
+def search_readmes(dir='Example Assignments'):
+	'''count the number of steps from the assignment readmes'''
+	data = {}
+	chdir(dir)
+	dirs = [d for d in listdir() if isdir(d)]
+	if 'Common Files' in dirs:
+		dirs.remove('Common Files')
+	for d in dirs:
+		chdir(d)
+		#check for subdirs
+		if not exists('README.md'):
+			sub = [s for s in listdir() if isdir(s)]
+			for s in sub:
+				chdir(s)
+				tag, min_commits = get_min_commits(s)
+				data[tag] = min_commits
+				chdir('..')
+		else:
+			tag, min_commits = get_min_commits(d)
+			data[tag] = min_commits
+		chdir('..')
+	chdir('..')
+	return data
+
+def get_min_commits(d):
+	'''extract tag from the given d(irectory)
+	count the number of steps in the README
+	min_commits is num_steps + 2 (auto generated by GitHub Classroom)'''
+	tag = d[:3]
+	if tag[0] != 'P': # projects are fine
+		try:
+			int(tag[0]) #try for regular assignment
+			tag = tag[:2]+'p'
+		except ValueError:
+			tag = 'sol' #and grab the interim assignment
+	with open('README.md','r') as f:
+		lines = f.readlines()
+	steps = []
+	for l in lines:
+		try:
+			int(l[0])#does the line start with a number?
+			steps.append(l)
+		except ValueError:
+			pass#do nothing if it doesn't
+	min_commits = len(steps)+2 #add 2 for the 2 commits added by GitHub Classroom
+	return tag, min_commits
 
 def execute(query):
 	'''connect to the default DATABASE_NAME and try to execute the provided query'''
@@ -39,6 +87,7 @@ def create():
 	assignments: tag(text, PK), total(int)
 		tag - The prefix tag for the assignment as made in GitHub Classroom
 		total - The total points the assignment is worth
+		min_commits - The minimum number of commits needed show your works for this assignment
 	scores: id(int PK), github(text, FK), tag(text, FK), earned(int)
 		id - Primary key used for sorting assignments so the reports display in the order that assignment were given
 		github - Foreign key to student's GitHub username
@@ -55,7 +104,8 @@ def create():
 	create_assignments_table='''
 	CREATE TABLE IF NOT EXISTS assignments (
 		tag TEXT PRIMARY KEY,
-		total INTEGER NOT NULL
+		total INTEGER NOT NULL,
+		min_commits INTEGER NOT NULL
 	);'''
 	execute(create_assignments_table)
 	create_scores_table='''
@@ -73,6 +123,7 @@ def create():
 	except FileNotFoundError as e:
 		print(f"Error while trying to load assignments:\n{e}")
 	if testing:
+		data = search_readmes()
 		for file in testing:
 			tag = file[:3]
 			try:
@@ -82,9 +133,10 @@ def create():
 			except ValueError:
 				#projects start with a letter and are worth more
 				points = 20
+			min_commits = data[tag]
 			enter_assignment=f'''
-			INSERT INTO assignments (tag, total)
-			VALUES ('{tag}', {points});'''
+			INSERT INTO assignments (tag, total, min_commits)
+			VALUES ('{tag}', {points}, {min_commits});'''
 			execute(enter_assignment)
 
 def select_tag():
